requirements.txt
annotated-types==0.7.0
anyio==4.9.0
fastapi==0.115.13
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
